Gather the election status types in election_status
election.hpp carried the extended status while the plain status lived alone
in election_status.hpp. Gather the type enum and both structs there, with
their object stream support and enum helpers in a matching source file, and
nest the composed status inside the extended report. election.hpp keeps only
the election machinery.
