+++ nano/node/active_elections.cpp
@@ -911,13 +911,3 @@ nano::stat::type nano::to_stat_type (nano::election_state state)
 	debug_assert (false);
 	return {};
 }
-
-std::string_view nano::to_string (nano::election_status_type type)
-{
-	return nano::enum_to_string (type);
-}
-
-nano::stat::detail nano::to_stat_detail (nano::election_status_type type)
-{
-	return nano::enum_convert<nano::stat::detail> (type);
-}
