+++ nano/node/election.hpp
@@ -42,18 +42,6 @@ struct election_actions final
 	bool cleanup{ false };
 };
 
-/** Extended election status with behavior and copies of the ballot state, all taken as one consistent snapshot */
-struct election_extended_status final
-{
-	nano::election_behavior behavior; // Scheduling behavior at the time of the snapshot
-	nano::election_status status; // The status summary
-	std::unordered_map<nano::account, nano::vote_info> votes; // All recorded votes by representative
-	std::unordered_map<nano::block_hash, std::shared_ptr<nano::block>> blocks; // All held blocks by hash
-	nano::tally_map tally; // Held blocks by tally position, zero-vote blocks absent
-
-	void operator() (nano::object_stream &) const;
-};
-
 enum class election_state
 {
 	passive, // only listening for incoming votes
